+++ example/lib/widgets/system_device_tile.dart
@@ -50,7 +50,7 @@ class _SystemDeviceTileState extends State<SystemDeviceTile> {
   Widget build(BuildContext context) {
     return ListTile(
       title: Text(widget.device.platformName),
-      subtitle: Text(widget.device.remoteId.toString()),
+      subtitle: Text(widget.device.remoteId.str),
       trailing: ElevatedButton(
         child: isConnected ? const Text('OPEN') : const Text('CONNECT'),
         onPressed: isConnected ? widget.onOpen : widget.onConnect,